implemented dependencies (needs tests)
